third_party/WebKit/LayoutTests/media/video-controls-mouse-events-captured-expected.txt

This tests that a mouse events on the controls will not be seen by the video element.

Also tests keyboard input.

EVENT(loadeddata)
END OF TEST
